Create keyboard
Added an empty div with the id of "keyboard."

Created the function createKeyboard.

Used let to create a string for all the letters of the alphabet.

Split this string up into individual letters, each with the id of "letter."

Joined them together with nothing.

Set the inner HTML of the keyboard to the letters.