Clean up
Replaced the empty div .keyboard with individual divs for each letter button.

Got rid of the function createKeyboard.

Put only two words in the array words.

Added the function pickLetter in all the letter buttons onclick.

This function finds the index of the chosen letter and adds it to the anonymous word if it exists.