Reset
Created the function reset and put it inside the reset button onclick.

This function sets the variable guessed back to an empty array.

It sets the variable wrong back to 0.

It sets the inner HTML of the span .wrong equal to the variable wrong.

It runs the functions pickWord and wordGuess.

It removes the disabled attribute from all the letter buttons.