Spaces
Added a space button where the chosenLetter is a hyphen.

Replaced all the hyphens in the variable answer with spaces so they will appear like that when the function wordGuess is run.

Created the variable finalWord to switch all the spaces back to hyphens so it can be compared with the variable answer to determine if you won the game.