Disable buttons
Added the class .letterHover to each letter button.

Used this class to achieve the hover and active effects.

Made the function pickWord remove this class from a button if it is pressed.

Made the functions youWon and youLost remove this class from all buttons.

Made the function reset add this class back to all buttons.