Pop up
Created the div #cover.

Fixed positioned it so it covers the entire screen.

Set the z-index to 1 so it is in front of everything else.

Made its background a transparent black.

Put the div #popUp inside of it.

Fixed positioned this to be in the center of the screen.

Used flexbox to distribute its contents along the center column axis.

Put an empty h2 in it.

Added a play again button.

Gave this button blue text and a blue border.

Set its background to turn transparent blue when hovered over.

Set it to scale down to 90% size when clicked on.

Added a close button.

Absolutely positioned it in the top right corner.

Set it to turn red with white text when hovered over.